refactor: persist map viewport in a jotai atom instead of sessionStorage
sessionStorage worked but was the wrong tool: it added string
serialization, a storage key + cleanup, and survived full page reloads
(which we don't need — the restore only needs to survive HomePage
unmounting across same-session route changes).

- new savedMapViewportAtom (module-scoped) written by ShelterCard on
  navigate and consumed (set to null) by HomePage's init effect
- HomePage initializes its controlled camera from the atom, so the Map
  mounts directly at the exact saved center + zoom
- mapViewportStorage.ts -> mapViewport.ts (keeps TMapViewport +
  mapViewportFromMap; the sessionStorage helpers are gone), and the
  SESSION_STORAGE_MAP_VIEWPORT key / sessionStorageKeys.ts are removed
- useRestoredMapViewport hook removed (replaced by the atom)
- tests updated: ShelterCard asserts on the atom; mapViewportFromMap
  tests kept; sessionStorage-based tests removed

Browser-verified: initial load, pan (no snap-back), search, and exact
viewport + count restore all work.
