+++ libs/react/shelter/src/lib/components/Map/utils/index.ts
@@ -1,12 +1,7 @@
 export { fitMapToPlaceViewport } from './fitMapToPlaceViewport';
 export { fromMapBounds } from './fromMapBounds';
 export { mapBoundsFromCenter } from './mapBoundsFromCenter';
-export {
-  consumeSavedMapViewport,
-  mapViewportFromMap,
-  peekSavedMapViewport,
-  saveMapViewport,
-} from './mapViewportStorage';
+export { mapViewportFromMap } from './mapViewport';
+export type { TMapViewport } from './mapViewport';
 export { toGoogleLatLng } from './toGoogleLatLng';
 export { toMapBounds } from './toMapBounds';
-export { useRestoredMapViewport } from './useRestoredMapViewport';
