+++ libs/react/shelter/src/lib/components/ShelterCard/ShelterCard.tsx
@@ -1,8 +1,10 @@
 import { mergeCss } from '@monorepo/react/shared';
 import { useMap } from '@vis.gl/react-google-maps';
+import { useSetAtom } from 'jotai';
 import { ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { TLatLng, mapViewportFromMap, saveMapViewport } from '../Map';
+import { savedMapViewportAtom } from '../../atoms';
+import { TLatLng, mapViewportFromMap } from '../Map';
 import { PrivateBadge } from '../PrivateBadge';
 import { DistanceAway } from './DistanceAway';
 import { ShelterCardHero } from './ShelterCardHero';
@@ -42,6 +44,7 @@ export function ShelterCard(props: TShelterCard) {
     footerClassName,
   } = props;
   const map = useMap();
+  const setSavedMapViewport = useSetAtom(savedMapViewportAtom);
 
   const navigate = useNavigate();
 
@@ -70,7 +73,7 @@ export function ShelterCard(props: TShelterCard) {
     const viewport = map ? mapViewportFromMap(map) : null;
 
     if (viewport) {
-      saveMapViewport(viewport);
+      setSavedMapViewport(viewport);
     }
 
     navigate(`/shelter/${id}`);
