+++ libs/react/shelter/src/lib/constants/index.ts
@@ -1,4 +1,3 @@
 export * from './appConstants';
 export * from './featureFlags';
 export * from './routePaths';
-export * from './sessionStorageKeys';
