--- libs/react/shelter/src/lib/constants/sessionStorageKeys.ts
@@ -1,2 +1,0 @@
-/** Key for the exact map viewport (center + zoom) saved before navigating to a shelter detail. */
-export const SESSION_STORAGE_MAP_VIEWPORT = 'savedMapViewport';
